+++ src/pages/404.tsx
@@ -5,7 +5,6 @@ import { NotFoundItemQuery } from 'graphql-types'
 import Layout from 'components/Layout'
 import Header from 'components/Header'
 import Navbar from 'components/Navbar'
-import Content from 'components/Content'
 import Footer from 'components/Footer'
 import NotFound from 'components/NotFound'
 
@@ -28,9 +27,7 @@ const NotFoundPage: FunctionComponent<NotFoundPageProps> = ({ data }) => (
   <Layout>
     <Header />
     <Navbar />
-    <Content>
-      <NotFound {...data} />
-    </Content>
+    <NotFound {...data} />
     <Footer />
   </Layout>
 )
