+++ src/pages/latest.tsx
@@ -5,7 +5,6 @@ import { LatestItemQuery } from 'graphql-types'
 import Layout from 'components/Layout'
 import Header from 'components/Header'
 import Navbar from 'components/Navbar'
-import Content from 'components/Content'
 import Footer from 'components/Footer'
 import Latest from 'components/Latest'
 
@@ -41,9 +40,7 @@ const LatestPage: FunctionComponent<LatestPageProps> = ({
   <Layout>
     <Header />
     <Navbar />
-    <Content>
-      <Latest items={items} />
-    </Content>
+    <Latest items={items} />
     <Footer />
   </Layout>
 )
