+++ src/pages/mail.tsx
@@ -3,7 +3,6 @@ import React, { FunctionComponent } from 'react'
 import Layout from 'components/Layout'
 import Header from 'components/Header'
 import Navbar from 'components/Navbar'
-import Content from 'components/Content'
 import Footer from 'components/Footer'
 import Mail from 'components/Mail'
 
@@ -11,9 +10,7 @@ const MailPage: FunctionComponent<PageProps> = () => (
   <Layout>
     <Header />
     <Navbar />
-    <Content>
-      <Mail />
-    </Content>
+    <Mail />
     <Footer />
   </Layout>
 )
