+++ src/templates/article.tsx
@@ -9,7 +9,6 @@ import Container from 'components/Container'
 import Metadata from 'components/Metadata'
 import Header from 'components/Header'
 import Navbar from 'components/Navbar'
-import Content from 'components/Content'
 import Footer from 'components/Footer'
 import Article from 'components/Article'
 
@@ -173,11 +172,9 @@ const ArticlePage: FunctionComponent<ArticlePageProps> = ({
       }} />
       <Header />
       <Navbar />
-      <Content sidebar={article.attributes.sidebar !== false}>
-        <Container>
-          <Article article={article} prev={prev} next={next} />
-        </Container>
-      </Content>
+      <Container sidebar={article.attributes.sidebar !== false}>
+        <Article article={article} prev={prev} next={next} />
+      </Container>
       <Footer />
     </Layout>
   )
